Remove rust-toolchain file
While our MSRV is 1.60.0, rust-analyzer runs into a bunch of problems
with this version where it can no longer expand proc macros. For
development, one should use their latest stable version that works with
whatever rust-analyzer or any other dev setup they have and not have to
be forced to use 1.60.

The downside is that it is easier to accidentally introduce code that
would require a later Rust version, but we should catch that with our
MSRV CI tests.