Retry on certain query failures for managed transactions
When using `Graph::run`, `Graph::run_on`, `Graph::execute`, or `Graph::execute_on`, the query might be retries for a while.
The logic is that all errors with the `Transient` error class as well as a few other error classes are considered retryable.
This catches errors during a leader election or when the transaction resources on the server (memory, handles, ...) are exhausted.
Retries happen with an exponential backoff until a retry delay exceeds 60s, at which point the query fails.

Transaction created with `Graph::start_txn` or `Graph::start_txn` are *not* retried automatically, only managed transactions are.