Simplify result stream by removing the RowOrSummary enum
The result summary can now only be returned from `finish`, which also consumes the stream.
The next_or_summary and *_items_* methods are removed.
This also simplifies the API for finish, the summary no longer has to be returned as an option.
Before it could have already been returned by a call to next_or_summary, which would result in a None being returned from finish.
Having the API like that is also closer to APIs from the product drivers